some method rest implmented

+++ BlaBlaLanguageWeb/src/Models/Establishment.java
@@ -11,8 +11,8 @@ public class Establishment{
     private User owner;
     public String ownerName;
     public String address;
-    private Integer longitude;
-    private Integer latitude;
+    private Double longitude;
+    private Double latitude;
     public Integer placesAvailable;
     public String  telephone;
 
@@ -48,19 +48,19 @@ public class Establishment{
         this.address = address;
     }
 
-    public Integer getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Integer longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
 
-    public Integer getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Integer latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
